+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
-## 0.13.3 or 0.14.0 (Unreleased)
+## 0.13.4 or 0.14.0 (Unreleased)
+
+## 0.13.3
 
 ### Important Changes